create model, adding ExecuteQuery to mongo
