+++ clients/storeClient.go
@@ -1,8 +1,11 @@
 package clients
 
+import "../model"
+
 type StoreClient interface {
 	Close()
-	Ping() error
+	ExecuteQuery(details model.QueryData) []byte
 	GetTimeLastEntryUser(groupId string) []byte
 	GetTimeLastEntryUserAndDevice(groupId, deviceId string) []byte
+	Ping() error
 }
